added cookbook class
